Remove error that breaks scheduler.
Credis does some array_walk thing that appears not to be recursive or something and the array parameter on one of the function calls in this lib has an array that causes an error and makes the scheduler not work. Removing that parameter doesn't seem to have any side-effects I can see... but fixes the error and makes the scheduler work again for me.